+++ CHANGES.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and
 this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 2018/07/04 [Unreleased][1]
+
+### Changed
+
+- *BREAKING*: relocated Delete-Button from `EditorBlock` to `ExampleBlockWrapper` => `Delete--Button` will not be included in `OyezEditor` itself anymore
+  - therefore `onBlockAction` function` relocated from </OyezEditor /> to <ExampleApp />
+
 ## [Unreleased][1]
 
 ### Added